refactor: extract catalog.ts and install-orchestrator.ts from consumer.ts
