+++ .changeset/refactor-composable-modules.md
@@ -1,0 +1,5 @@
+---
+"pkglab": patch
+---
+
+refactor: extract catalog.ts and install-orchestrator.ts from consumer.ts and pub.ts
