+++ src/commands/add.ts
@@ -5,6 +5,7 @@ import type { VersionEntry, CatalogFormat } from '../lib/consumer';
 import type { pkglabConfig, RepoState } from '../types';
 
 import { getPositionalArgs, normalizeScope } from '../lib/args';
+import { findCatalogRoot, findCatalogEntry, loadCatalogData } from '../lib/catalog';
 import { c } from '../lib/color';
 import { loadConfig } from '../lib/config';
 import {
@@ -12,9 +13,6 @@ import {
   injectPreCommitHook,
   ensureNpmrcForActiveRepos,
   installWithVersionUpdates,
-  findCatalogRoot,
-  findCatalogEntry,
-  loadCatalogData,
 } from '../lib/consumer';
 import { ensureDaemonRunning } from '../lib/daemon';
 import { CommandError, SilentExitError } from '../lib/errors';
